Changed voteBonus to support multiple types of bot awesoming
The voteBonus flag tells the bot when it can awesome a song. Currently,
these modes are supported:
NONE: The bot will never awesome
VOTE: The bot awesomes when a vote threshold is met (# of awesomes)
CHAT: The bot awesomes when enough people say a bonus phrase
AUTO: The bot auto-awesomes
(Note: Turntable.fm does not allow bots to auto-awesome. If you set
your bot to auto-awesome, it is likely that the bot will be deleted.
The implementation is included as a reference only.)