Added SparkleConfig.app
SparkleConfig.app is a quick/dirty way for OSX users to generate a
config file. I'm planning on making it much nicer in the future, but
this is a nice alternative to manually fighting through a JSON file.
Make sure all information is correct before generating the output file
- there's no bounds checking yet. It currently saves to your Home
folder (that will change in the future as well).